requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
datasets==3.0.2
dill==0.3.8
filelock==3.16.1
flwr==1.12.0
flwr-datasets==0.4.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.67.0
huggingface-hub==0.26.1
idna==3.10
importlib_resources==6.4.5
iterators==0.0.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==16.1.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ray==2.10.0
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tomli==2.0.2
tomli_w==1.1.0
torch==2.5.0
torchinfo==1.8.0
torchvision==0.20.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
